+++ src/mspparser.hpp
@@ -626,6 +626,24 @@ namespace hf {
                         serialize8(_checksum);
                         } break;
 
+                    case 25:
+                    {
+                        uint8_t mosquitoVersion = 0;
+                        handle_MOSQUITO_VERSION_Request(mosquitoVersion);
+                        prepareToSendBytes(1);
+                        sendByte(mosquitoVersion);
+                        serialize8(_checksum);
+                        } break;
+
+                    case 26:
+                    {
+                        uint8_t hasPositionBoard = 0;
+                        handle_POSITION_BOARD_Request(hasPositionBoard);
+                        prepareToSendBytes(1);
+                        sendByte(hasPositionBoard);
+                        serialize8(_checksum);
+                        } break;
+
                     case 30:
                     {
                         uint8_t flag = 0;
@@ -885,6 +903,18 @@ namespace hf {
                     {
                         uint8_t flag = getArgument(0);
                         handle_WP_MISSION_FLAG_Data(flag);
+                        } break;
+
+                    case 25:
+                    {
+                        uint8_t mosquitoVersion = getArgument(0);
+                        handle_MOSQUITO_VERSION_Data(mosquitoVersion);
+                        } break;
+
+                    case 26:
+                    {
+                        uint8_t hasPositionBoard = getArgument(0);
+                        handle_POSITION_BOARD_Data(hasPositionBoard);
                         } break;
 
                     case 30:
@@ -1242,6 +1272,26 @@ namespace hf {
             virtual void handle_WP_MISSION_FLAG_Data(uint8_t & flag)
             {
                 (void)flag;
+            }
+
+            virtual void handle_MOSQUITO_VERSION_Request(uint8_t & mosquitoVersion)
+            {
+                (void)mosquitoVersion;
+            }
+
+            virtual void handle_MOSQUITO_VERSION_Data(uint8_t & mosquitoVersion)
+            {
+                (void)mosquitoVersion;
+            }
+
+            virtual void handle_POSITION_BOARD_Request(uint8_t & hasPositionBoard)
+            {
+                (void)hasPositionBoard;
+            }
+
+            virtual void handle_POSITION_BOARD_Data(uint8_t & hasPositionBoard)
+            {
+                (void)hasPositionBoard;
             }
 
             virtual void handle_WP_MISSION_BEGIN_Request(uint8_t & flag)
@@ -2014,6 +2064,60 @@ namespace hf {
                 bytes[4] = 23;
 
                 memcpy(&bytes[5], &flag, sizeof(uint8_t));
+
+                bytes[6] = CRC8(&bytes[3], 3);
+
+                return 7;
+            }
+
+            static uint8_t serialize_MOSQUITO_VERSION_Request(uint8_t bytes[])
+            {
+                bytes[0] = 36;
+                bytes[1] = 77;
+                bytes[2] = 60;
+                bytes[3] = 0;
+                bytes[4] = 25;
+                bytes[5] = 25;
+
+                return 6;
+            }
+
+            static uint8_t serialize_MOSQUITO_VERSION(uint8_t bytes[], uint8_t  mosquitoVersion)
+            {
+                bytes[0] = 36;
+                bytes[1] = 77;
+                bytes[2] = 62;
+                bytes[3] = 1;
+                bytes[4] = 25;
+
+                memcpy(&bytes[5], &mosquitoVersion, sizeof(uint8_t));
+
+                bytes[6] = CRC8(&bytes[3], 3);
+
+                return 7;
+            }
+
+            static uint8_t serialize_POSITION_BOARD_Request(uint8_t bytes[])
+            {
+                bytes[0] = 36;
+                bytes[1] = 77;
+                bytes[2] = 60;
+                bytes[3] = 0;
+                bytes[4] = 26;
+                bytes[5] = 26;
+
+                return 6;
+            }
+
+            static uint8_t serialize_POSITION_BOARD(uint8_t bytes[], uint8_t  hasPositionBoard)
+            {
+                bytes[0] = 36;
+                bytes[1] = 77;
+                bytes[2] = 62;
+                bytes[3] = 1;
+                bytes[4] = 26;
+
+                memcpy(&bytes[5], &hasPositionBoard, sizeof(uint8_t));
 
                 bytes[6] = CRC8(&bytes[3], 3);