+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 * #37 Hackflight Parametrization
 * #46 Implement handler for getting Motor values via MSP
+* #47 MSP getters for Mosquito version and presence of Position Board
 
 ### Changed